breast_cancer_ultrasound/requirements.txt
tensorflow
numpy
Pillow
opencv-python-headless
streamlit
